requirements.txt
sounddevice
flask
requests
numpy
whisper-openai
time